adicionando metodos Post e Delete

+++ src/app/grupo/grupo.component.ts
@@ -1,14 +1,15 @@
 import { Component } from '@angular/core';
-import { Grupo } from '../model/grupo';
+import { GrupoInterface } from '../model/grupo';
 import { GrupoService } from '../services/grupo.service';
 import { MatDialog } from '@angular/material/dialog';
 import { GrupoDialogInserindoComponent } from './grupo-dialog-inserindo/grupo-dialog-inserindo.component';
 import { GrupoDialogEditandoComponent } from './grupo-dialog-editando/grupo-dialog-editando.component';
 import { GrupoDialogExcluindoComponent } from './grupo-dialog-excluindo/grupo-dialog-excluindo.component';
-import { Subgrupo } from '../model/subgrupo';
-import { Produto } from '../model/produto';
+import { SubgrupoInterface } from '../model/subgrupo';
+import { ProdutoInterface } from '../model/produto';
 import { SubgrupoService } from '../services/subgrupo.service';
 import { ProdutoService } from '../services/produto.service';
+import { Grupo } from '../grupo';
 @Component({
   selector: 'app-grupo',
   templateUrl: './grupo.component.html',
@@ -16,40 +17,44 @@ import { ProdutoService } from '../services/produto.service';
 })
 export class GrupoComponent {
 
-  grupo: Grupo[] = [];
-  subgrupo: Subgrupo[] = [];
-  produto: Produto[] = [];
+  grupo: GrupoInterface[] = [];
+  subgrupo: SubgrupoInterface[] = [];
+
 
   displayedColumns = ['botoes', 'nome', 'descricao'];
 
-  constructor(public dialog: MatDialog, private grupoService: GrupoService, private subgrupoService: SubgrupoService, private produtoService: ProdutoService) {
+  constructor(public dialog: MatDialog, private grupoService: GrupoService, private subgrupoService: SubgrupoService) {
     this.getGroups();
   }
 
   getGroups() {
     this.grupoService.getGroups().subscribe(grupo => this.grupo = grupo);
     this.subgrupoService.getSubgroups().subscribe(subgrupo => this.subgrupo = subgrupo);
-    this.produtoService.getProducts().subscribe(produto => this.produto = produto);
   }
 
   deleteGroup(idGroup: Number) {
-    const subgruposParaDeletar = this.subgrupo.filter(subgrupo => subgrupo.idProdutoGrupo === idGroup);
-    subgruposParaDeletar.forEach(subgrupo => {
-      const produtosParaDeletar = this.produto.filter(produto => produto.idProdutoSubgrupo === subgrupo.id);
-      produtosParaDeletar.forEach(produto => {
-        this.produtoService.deleteProducts(produto.id!).subscribe(_ => this.getGroups());
-      });
+    const relatedItens = this.subgrupo.filter(subgrupo => subgrupo.idProdutoGrupo === idGroup);
+    if (relatedItens.length == 0) {
+      this.grupoService.deleteGroup(idGroup).subscribe(_ => this.getGroups());
+    } else {
+      alert(`Ação bloqueada\nExistem ${relatedItens.length} Subgrupos dentro deste Grupo no momento`);
+    }
+  }
 
-      this.subgrupoService.deleteSubgroup(subgrupo.id!).subscribe(_ => this.getGroups());
-    });
-    this.grupoService.deleteGroup(idGroup).subscribe(_ => this.getGroups());
+  postGroup(grupo: Grupo) {
+    this.grupoService.postGroup(grupo).subscribe(_ => this.getGroups());
   }
 
 
   openDialog(enterAnimationDuration: string, exitAnimationDuration: string): void {
-    this.dialog.open(GrupoDialogInserindoComponent, {
+    const dialogRef = this.dialog.open(GrupoDialogInserindoComponent, {
       width: '100%',
     });
+    dialogRef.afterClosed().subscribe(newGroup => {
+      if (newGroup) {
+        this.postGroup(newGroup);
+      }
+    })
   }
 
   openDialogEdit(enterAnimationDuration: string, exitAnimationDuration: string): void {
